+++ compile/py/compiler.go
@@ -1235,6 +1235,8 @@ func (c *Compiler) compileCallExpr(call *parser.CallExpr) (string, error) {
 	case "avg":
 		c.use("_avg")
 		return fmt.Sprintf("_avg(%s)", argStr), nil
+	case "scan":
+		return "input().strip()", nil
 	case "eval":
 		return fmt.Sprintf("eval(%s)", argStr), nil
 	default: